Add files via upload
In function call when values of variable is pass to the calling function then the value of actual parameter is copied in formal parameter. then it is called call by value.(In this call by value function, function are called by variable).When the function is called the separate copy of actual parameter is created in memory.And, the value of original variable is given to this variable. So, if any changes made in the value it will not reflect to the original variable.